backend implementado

+++ models/services.go
@@ -4,12 +4,11 @@ import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-/*Usuario es el modelo de usuario de la base de MongoDB */
 type Service struct {
 	ID          primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	Servicename string             `bson:"Servicename" json:"Servicename,omitempty"` //user-name
 	Description string             `bson:"description" json:"description,omitempty"`
-	Price       int                `bson:"price" json:"price,omitempty"`
+	Price       float64            `bson:"price" json:"price,omitempty"`
 	Available   bool               `bson:"available" json:"available"`
 	Category    string             `bson:"category" json:"category,omitempty"`
 }
